Collapse all retrieve-and-print executors into one.
